Working example of creating Pandas df from different file types
Small refactor to make this its own function. Idea is for user to be able to
load their own data into a df if they need to and then use the rest of our
functions from there.

Still needs tests, logging, proper errors, and a way to deal with SMILES that
don't parse.